+++ Src/DependencyCollector/Net45.Tests/DependencyTrackingTelemetryModuleHttpTest.cs
@@ -35,12 +35,18 @@
         private StubTelemetryChannel channel;
         private TelemetryConfiguration config;
         private List<DependencyTelemetry> sentTelemetry;
+        private object request;
+        private object response;
+        private object responseHeaders;
 
         [TestInitialize]
         public void Initialize()
         {
             ServicePointManager.DefaultConnectionLimit = 1000;
             this.sentTelemetry = new List<DependencyTelemetry>();
+            this.request = null;
+            this.response = null;
+            this.responseHeaders = null;
 
             this.channel = new StubTelemetryChannel
             {
@@ -51,6 +57,9 @@
                     if (depTelemetry != null)
                     {
                         this.sentTelemetry.Add(depTelemetry);
+                        depTelemetry.TryGetOperationDetail(RemoteDependencyConstants.HttpRequestOperationDetailName, out this.request);
+                        depTelemetry.TryGetOperationDetail(RemoteDependencyConstants.HttpResponseOperationDetailName, out this.response);
+                        depTelemetry.TryGetOperationDetail(RemoteDependencyConstants.HttpResponseOperationDetailName, out this.responseHeaders);
                     }
                 },
                 EndpointAddress = FakeProfileApiEndpoint
@@ -382,7 +391,7 @@
                     }
                 }
 
-                this.ValidateTelemetry(enableDiagnosticSource, this.sentTelemetry.Single(), new Uri(url), request, statusCode >= 200 && statusCode < 300, statusCode.ToString(CultureInfo.InvariantCulture), injectLegacyHeaders);
+                this.ValidateTelemetry(enableDiagnosticSource, this.sentTelemetry.Single(), new Uri(url), request, statusCode >= 200 && statusCode < 300, statusCode.ToString(CultureInfo.InvariantCulture), expectLegacyHeaders: injectLegacyHeaders);
             }
         }
 
@@ -412,7 +421,7 @@
                     }
                 }
 
-                this.ValidateTelemetry(true, this.sentTelemetry.Single(), new Uri(url), null, statusCode >= 200 && statusCode < 300, statusCode.ToString(CultureInfo.InvariantCulture));
+                this.ValidateTelemetry(true, this.sentTelemetry.Single(), new Uri(url), null, statusCode >= 200 && statusCode < 300, statusCode.ToString(CultureInfo.InvariantCulture), responseExpected: contentLength != 0);
             }
         }
 
@@ -468,7 +477,7 @@
                 }
 
                 Assert.AreEqual(2, this.sentTelemetry.Count);
-                this.ValidateTelemetry(true, this.sentTelemetry.Last(), new Uri(url), null, statusCode >= 200 && statusCode < 300, statusCode.ToString(CultureInfo.InvariantCulture));
+                this.ValidateTelemetry(true, this.sentTelemetry.Last(), new Uri(url), null, statusCode >= 200 && statusCode < 300, statusCode.ToString(CultureInfo.InvariantCulture), responseExpected: false);
             }
         }
 
@@ -498,7 +507,7 @@
                     await httpClient.GetAsync(url, cts.Token).ContinueWith(t => { });
                 }
 
-                this.ValidateTelemetry(enableDiagnosticSource, this.sentTelemetry.Single(), new Uri(url), null, false, string.Empty);
+                this.ValidateTelemetry(enableDiagnosticSource, this.sentTelemetry.Single(), new Uri(url), null, false, string.Empty, responseExpected: false);
             }
         }
 
@@ -515,7 +524,7 @@
                     // here the start of dependency is tracked with HttpDesktopDiagnosticSourceListener, 
                     // so the expected SDK version should have DiagnosticSource 'rdddsd' prefix. 
                     // however the end is tracked by FrameworkHttpEventListener
-                    this.ValidateTelemetry(true, this.sentTelemetry.Single(), url, null, false, string.Empty);
+                    this.ValidateTelemetry(true, this.sentTelemetry.Single(), url, null, false, string.Empty, responseExpected: false);
                 }
                 else
                 {
@@ -526,7 +535,7 @@
             }
         }
 
-        private void ValidateTelemetry(bool diagnosticSource, DependencyTelemetry item, Uri url, WebRequest request, bool success, string resultCode, bool expectLegacyHeaders = false)
+        private void ValidateTelemetry(bool diagnosticSource, DependencyTelemetry item, Uri url, WebRequest request, bool success, string resultCode, bool responseExpected = true, bool headersExpected = false, bool expectLegacyHeaders = false)
         {
             Assert.AreEqual(url, item.Data);
 
@@ -554,6 +563,30 @@
 
             Assert.AreEqual(Activity.Current?.Id, item.Context.Operation.ParentId);
             Assert.IsTrue(item.Id.StartsWith('|' + item.Context.Operation.Id + '.'));
+
+            // Validate the http request was captured
+            if (diagnosticSource)
+            {
+                Assert.IsNotNull(this.request, "Http request was not found within the operation details.");
+                var webRequest = this.request as WebRequest;
+                Assert.IsNotNull(webRequest, "Http request was not the expected type.");
+            }
+
+            // If expected -- validate the response was captured
+            if (diagnosticSource && responseExpected)
+            {
+                Assert.IsNotNull(this.response, "Http response was not found within the operation details.");
+                var webResponse = this.response as WebResponse;
+                Assert.IsNotNull(webResponse, "Http response was not the expected type.");
+            }
+
+            // If expected -- validate the headers were captured
+            if (diagnosticSource && headersExpected)
+            {
+                Assert.IsNotNull(this.responseHeaders, "Http response headers were not found within the operation details.");
+                var headers = this.responseHeaders as WebHeaderCollection;
+                Assert.IsNotNull(headers, "Http response headers were not the expected type.");
+            }
 
             if (diagnosticSource)
             {