+++ Src/DependencyCollector/Shared.Tests/Implementation/DependencyCollectorDiagnosticListenerTests.Netstandard16.cs
@@ -29,7 +29,10 @@ namespace Microsoft.ApplicationInsights.Tests
         private const string HttpOkResultCode = "200";
         private const string NotFoundResultCode = "404";
 
-        private readonly List<ITelemetry> sentTelemetry = new List<ITelemetry>();
+        private List<ITelemetry> sentTelemetry;
+        private object request;
+        private object response;
+        private object responseHeaders;
 
         private TelemetryConfiguration configuration;
         private string testInstrumentationKey1 = nameof(testInstrumentationKey1);
@@ -44,10 +47,27 @@ namespace Microsoft.ApplicationInsights.Tests
         [TestInitialize]
         public void Initialize()
         {
+            this.sentTelemetry = new List<ITelemetry>();
+            this.request = null;
+            this.response = null;
+            this.responseHeaders = null;
+
             this.telemetryChannel = new StubTelemetryChannel()
             {
                 EndpointAddress = "https://endpointaddress",
-                OnSend = this.sentTelemetry.Add
+                OnSend = telemetry =>
+                {
+                    this.sentTelemetry.Add(telemetry);
+
+                    // The correlation id lookup service also makes http call, just make sure we skip that
+                    DependencyTelemetry depTelemetry = telemetry as DependencyTelemetry;
+                    if (depTelemetry != null)
+                    {
+                        depTelemetry.TryGetOperationDetail(RemoteDependencyConstants.HttpRequestOperationDetailName, out this.request);
+                        depTelemetry.TryGetOperationDetail(RemoteDependencyConstants.HttpResponseOperationDetailName, out this.response);
+                        depTelemetry.TryGetOperationDetail(RemoteDependencyConstants.HttpResponseHeadersOperationDetailName, out this.responseHeaders);
+                    }
+                },
             };
 
             this.testInstrumentationKey1 = Guid.NewGuid().ToString();
@@ -314,6 +334,9 @@ namespace Microsoft.ApplicationInsights.Tests
             string expectedVersion =
                 SdkVersionHelper.GetExpectedSdkVersion(typeof(DependencyTrackingTelemetryModule), prefix: "rdddsc:");
             Assert.AreEqual(expectedVersion, telemetry.Context.GetInternalContext().SdkVersion);
+
+            // Check the operation details
+            this.ValidateOperationDetails(telemetry);
         }
 
         /// <summary>
@@ -349,6 +372,9 @@ namespace Microsoft.ApplicationInsights.Tests
             Assert.AreEqual(RequestUrl, telemetry.Target);
             Assert.AreEqual(NotFoundResultCode, telemetry.ResultCode);
             Assert.AreEqual(false, telemetry.Success);
+
+            // Check the operation details
+            this.ValidateOperationDetails(telemetry);
         }
 
         /// <summary>
@@ -385,6 +411,9 @@ namespace Microsoft.ApplicationInsights.Tests
             Assert.AreEqual(RequestUrl, telemetry.Target);
             Assert.AreEqual(HttpOkResultCode, telemetry.ResultCode);
             Assert.AreEqual(true, telemetry.Success, "response was not successful");
+
+            // Check the operation details
+            this.ValidateOperationDetails(telemetry);
         }
 
         /// <summary>
@@ -420,6 +449,9 @@ namespace Microsoft.ApplicationInsights.Tests
             Assert.AreEqual(RequestUrl, telemetry.Target);
             Assert.AreEqual(NotFoundResultCode, telemetry.ResultCode);
             Assert.AreEqual(false, telemetry.Success);
+
+            // Check the operation details
+            this.ValidateOperationDetails(telemetry);
         }
 
         /// <summary>
@@ -456,6 +488,9 @@ namespace Microsoft.ApplicationInsights.Tests
             Assert.AreEqual(GetApplicationInsightsTarget(targetApplicationId), telemetry.Target);
             Assert.AreEqual(HttpOkResultCode, telemetry.ResultCode);
             Assert.AreEqual(true, telemetry.Success);
+
+            // Check the operation details
+            this.ValidateOperationDetails(telemetry);
         }
 
         /// <summary>
@@ -492,6 +527,9 @@ namespace Microsoft.ApplicationInsights.Tests
             Assert.AreEqual(GetApplicationInsightsTarget(targetApplicationId), telemetry.Target);
             Assert.AreEqual(NotFoundResultCode, telemetry.ResultCode);
             Assert.AreEqual(false, telemetry.Success);
+
+            // Check the operation details
+            this.ValidateOperationDetails(telemetry);
         }
 
         /// <summary>
@@ -534,6 +572,9 @@ namespace Microsoft.ApplicationInsights.Tests
             Assert.AreEqual(parentActivity.RootId, telemetry.Context.Operation.Id);
             Assert.AreEqual(parentActivity.Id, telemetry.Context.Operation.ParentId);
 
+            // Check the operation details
+            this.ValidateOperationDetails(telemetry);
+
             parentActivity.Stop();
         }
 
@@ -550,6 +591,22 @@ namespace Microsoft.ApplicationInsights.Tests
         private static string GetRequestContextKeyValue(HttpRequestMessage request, string keyName)
         {
             return HttpHeadersUtilities.GetRequestContextKeyValue(request.Headers, keyName);
+        }
+
+        private void ValidateOperationDetails(DependencyTelemetry telemetry, bool responseExpected = true)
+        {
+            Assert.IsNotNull(this.request, "Request was not present and expected.");
+            Assert.IsNotNull(this.request as HttpRequestMessage, "Request was not the expected type.");
+            Assert.IsNull(this.responseHeaders, "Response headers were present and not expected.");
+            if (responseExpected)
+            {
+                Assert.IsNotNull(this.response, "Response was not present and expected.");
+                Assert.IsNotNull(this.response as HttpResponseMessage, "Response was not the expected type.");
+            }
+            else
+            {
+                Assert.IsNull(this.response, "Response was present and not expected.");
+            }
         }
     }
 }