+++ Src/DependencyCollector/Shared.Tests/Implementation/DependencyCollectorDiagnosticListenerTests.Netstandard20.cs
@@ -106,6 +106,9 @@ namespace Microsoft.ApplicationInsights.Tests
             string expectedVersion =
                 SdkVersionHelper.GetExpectedSdkVersion(typeof(DependencyTrackingTelemetryModule), prefix: "rdddsc:");
             Assert.AreEqual(expectedVersion, telemetry.Context.GetInternalContext().SdkVersion);
+
+            // Check the operation details
+            this.ValidateOperationDetails(telemetry);
         }
 
         /// <summary>
@@ -126,6 +129,9 @@ namespace Microsoft.ApplicationInsights.Tests
 
             Assert.AreEqual("Canceled", telemetry.ResultCode);
             Assert.AreEqual(false, telemetry.Success);
+
+            // Check the operation details
+            this.ValidateOperationDetails(telemetry, responseExpected: false);
         }
 
         /// <summary>
@@ -146,6 +152,9 @@ namespace Microsoft.ApplicationInsights.Tests
 
             Assert.AreEqual("Faulted", telemetry.ResultCode);
             Assert.AreEqual(false, telemetry.Success);
+
+            // Check the operation details
+            this.ValidateOperationDetails(telemetry, responseExpected: false);
         }
 
         /// <summary>
@@ -172,6 +181,9 @@ namespace Microsoft.ApplicationInsights.Tests
             Assert.AreEqual(exceptionTelemetry.Context.Operation.Id, dependencyTelemetry.Context.Operation.Id);
             Assert.AreEqual(exceptionTelemetry.Context.Operation.ParentId, dependencyTelemetry.Id);
             Assert.AreEqual("The server name or address could not be resolved", dependencyTelemetry.Context.Properties["Error"]);
+
+            // Check the operation details
+            this.ValidateOperationDetails(dependencyTelemetry, responseExpected: false);
         }
 
         /// <summary>