+++ Src/DependencyCollector/Shared.Tests/Implementation/ProfilerHttpProcessingTest.cs
@@ -35,7 +35,10 @@
         private TelemetryConfiguration configuration;
         private Uri testUrl = new Uri("http://www.microsoft.com/");
         private Uri testUrlNonStandardPort = new Uri("http://www.microsoft.com:911/");
-        private List<ITelemetry> sendItems = new List<ITelemetry>();
+        private List<ITelemetry> sendItems;
+        private object request;
+        private object response;
+        private object responseHeaders;
         private int sleepTimeMsecBetweenBeginAndEnd = 100;
         private Exception ex = new Exception();
         private ProfilerHttpProcessing httpProcessingProfiler;
@@ -46,9 +49,29 @@
         [TestInitialize]
         public void TestInitialize()
         {
+            this.sendItems = new List<ITelemetry>();
+            this.request = null;
+            this.response = null;
+            this.responseHeaders = null;
+
             this.configuration = new TelemetryConfiguration()
             {
-                TelemetryChannel = new StubTelemetryChannel { OnSend = item => this.sendItems.Add(item) },
+                TelemetryChannel = new StubTelemetryChannel
+                {
+                    OnSend = telemetry =>
+                    {
+                        this.sendItems.Add(telemetry);
+
+                        // The correlation id lookup service also makes http call, just make sure we skip that
+                        DependencyTelemetry depTelemetry = telemetry as DependencyTelemetry;
+                        if (depTelemetry != null)
+                        {
+                            depTelemetry.TryGetOperationDetail(RemoteDependencyConstants.HttpRequestOperationDetailName, out this.request);
+                            depTelemetry.TryGetOperationDetail(RemoteDependencyConstants.HttpResponseOperationDetailName, out this.response);
+                            depTelemetry.TryGetOperationDetail(RemoteDependencyConstants.HttpResponseHeadersOperationDetailName, out this.responseHeaders);
+                        }
+                    },
+                },
                 InstrumentationKey = TestInstrumentationKey,
                 ApplicationIdProvider = new MockApplicationIdProvider(TestInstrumentationKey, TestApplicationId)
             };
@@ -83,11 +106,11 @@
         [Owner("cithomas")]
         [TestCategory("CVT")]
         public void RddTestHttpProcessingProfilerOnBeginForGetResponse()
-        {                        
+        {
             var request = WebRequest.Create(this.testUrl);
             DependencyTelemetry operationReturned = (DependencyTelemetry)this.httpProcessingProfiler.OnBeginForGetResponse(request);
             Assert.IsNull(operationReturned, "Operation returned should be null as all context is maintained internally");
-            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");         
+            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");
         }
 
         /// <summary>
@@ -98,20 +121,20 @@
         [Owner("cithomas")]
         [TestCategory("CVT")]
         public void RddTestHttpProcessingProfilerOnEndForGetResponse()
-        {            
+        {
             var request = WebRequest.Create(this.testUrl);
             var returnObjectPassed = TestUtils.GenerateHttpWebResponse(HttpStatusCode.OK);
             this.httpProcessingProfiler.OnBeginForGetResponse(request);
             Stopwatch stopwatch = new Stopwatch();
             stopwatch.Start();
             Thread.Sleep(this.sleepTimeMsecBetweenBeginAndEnd);
-            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");         
+            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");
             var objectReturned = this.httpProcessingProfiler.OnEndForGetResponse(null, returnObjectPassed, request);
             stopwatch.Stop();
 
             Assert.AreSame(returnObjectPassed, objectReturned, "Object returned from OnEndForGetResponse processor is not the same as expected return object");
             Assert.AreEqual(1, this.sendItems.Count, "Only one telemetry item should be sent");
-            ValidateTelemetryPacket(this.sendItems[0] as DependencyTelemetry, this.testUrl, RemoteDependencyConstants.HTTP, true, stopwatch.Elapsed.TotalMilliseconds, "200");
+            this.ValidateTelemetryPacket(this.sendItems[0] as DependencyTelemetry, this.testUrl, RemoteDependencyConstants.HTTP, true, stopwatch.Elapsed.TotalMilliseconds, "200");
         }
 
         /// <summary>
@@ -327,12 +350,12 @@
             this.httpProcessingProfiler.OnBeginForGetResponse(request);
             Thread.Sleep(this.sleepTimeMsecBetweenBeginAndEnd);
             Exception exc = new Exception();
-            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");         
+            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");
             this.httpProcessingProfiler.OnExceptionForGetResponse(null, exc, request);
             stopwatch.Stop();
 
             Assert.AreEqual(1, this.sendItems.Count, "Only one telemetry item should be sent");
-            ValidateTelemetryPacket(this.sendItems[0] as DependencyTelemetry, this.testUrl, RemoteDependencyConstants.HTTP, false, stopwatch.Elapsed.TotalMilliseconds, string.Empty);
+            this.ValidateTelemetryPacket(this.sendItems[0] as DependencyTelemetry, this.testUrl, RemoteDependencyConstants.HTTP, false, stopwatch.Elapsed.TotalMilliseconds, string.Empty, responseExpected: false);
         }
 
         /// <summary>
@@ -356,7 +379,7 @@
             stopwatch.Stop();
 
             Assert.AreEqual(1, this.sendItems.Count, "Only one telemetry item should be sent");
-            ValidateTelemetryPacket(this.sendItems[0] as DependencyTelemetry, this.testUrl, RemoteDependencyConstants.HTTP, false, stopwatch.Elapsed.TotalMilliseconds, "404");
+            this.ValidateTelemetryPacket(this.sendItems[0] as DependencyTelemetry, this.testUrl, RemoteDependencyConstants.HTTP, false, stopwatch.Elapsed.TotalMilliseconds, "404");
         }
 
         /// <summary>
@@ -400,7 +423,7 @@
                 DependencyTelemetry operationReturned = (DependencyTelemetry)this.httpProcessingProfiler.OnBeginForGetResponse(request);
                 var objectReturned = this.httpProcessingProfiler.OnEndForGetResponse(null, returnObjectPassed, null);
                 Assert.AreSame(returnObjectPassed, objectReturned, "Object returned from OnEndForGetResponse processor is not the same as expected return object");
-                Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");         
+                Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");
                 
                 var message = listener.Messages.First(item => item.EventId == 14);
                 Assert.IsNotNull(message);  
@@ -423,7 +446,7 @@
             var request = WebRequest.Create(this.testUrl);
             DependencyTelemetry operationReturned = (DependencyTelemetry)this.httpProcessingProfiler.OnBeginForGetRequestStream(request, null);
             Assert.IsNull(operationReturned, "Operation returned should be null as all context is maintained internally");
-            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");         
+            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");
         }
         
         /// <summary>
@@ -441,13 +464,13 @@
             stopwatch.Start();
             this.httpProcessingProfiler.OnBeginForGetResponse(request);
             Thread.Sleep(this.sleepTimeMsecBetweenBeginAndEnd);
-            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");         
+            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");
             Exception exc = new Exception();
             this.httpProcessingProfiler.OnExceptionForGetResponse(null, exc, request);
             stopwatch.Stop();
 
             Assert.AreEqual(1, this.sendItems.Count, "Only one telemetry item should be sent");
-            ValidateTelemetryPacket(this.sendItems[0] as DependencyTelemetry, this.testUrl, RemoteDependencyConstants.HTTP, false, stopwatch.Elapsed.TotalMilliseconds, string.Empty);
+            this.ValidateTelemetryPacket(this.sendItems[0] as DependencyTelemetry, this.testUrl, RemoteDependencyConstants.HTTP, false, stopwatch.Elapsed.TotalMilliseconds, string.Empty, responseExpected: false);
         }
 
         #endregion //GetRequestStream
@@ -466,7 +489,7 @@
             var request = WebRequest.Create(this.testUrl);
             DependencyTelemetry operationReturned = (DependencyTelemetry)this.httpProcessingProfiler.OnBeginForBeginGetResponse(request, null, null);
             Assert.IsNull(operationReturned, "For async methods, operation returned should be null as correlation is done internally using WeakTables.");
-            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");         
+            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");
         }
 
         /// <summary>
@@ -484,13 +507,13 @@
             stopwatch.Start();
             DependencyTelemetry operationReturned = (DependencyTelemetry)this.httpProcessingProfiler.OnBeginForBeginGetResponse(request, null, null);
             Thread.Sleep(this.sleepTimeMsecBetweenBeginAndEnd);
-            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");         
+            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");
             var objectReturned = this.httpProcessingProfiler.OnEndForEndGetResponse(operationReturned, returnObjectPassed, request, null);
             stopwatch.Stop();
 
             Assert.AreSame(returnObjectPassed, objectReturned, "Object returned from OnEndForEndGetResponse processor is not the same as expected return object");
             Assert.AreEqual(1, this.sendItems.Count, "Only one telemetry item should be sent");
-            ValidateTelemetryPacket(this.sendItems[0] as DependencyTelemetry, this.testUrl, RemoteDependencyConstants.HTTP, true, stopwatch.Elapsed.TotalMilliseconds, "200");
+            this.ValidateTelemetryPacket(this.sendItems[0] as DependencyTelemetry, this.testUrl, RemoteDependencyConstants.HTTP, true, stopwatch.Elapsed.TotalMilliseconds, "200");
         }
 
         /// <summary>
@@ -514,7 +537,7 @@
 
             Assert.AreSame(returnObjectPassed, objectReturned, "Object returned from OnEndForEndGetResponse processor is not the same as expected return object");
             Assert.AreEqual(1, this.sendItems.Count, "Only one telemetry item should be sent");
-            ValidateTelemetryPacket(this.sendItems[0] as DependencyTelemetry, this.testUrl, RemoteDependencyConstants.HTTP, false, stopwatch.Elapsed.TotalMilliseconds, string.Empty);
+            this.ValidateTelemetryPacket(this.sendItems[0] as DependencyTelemetry, this.testUrl, RemoteDependencyConstants.HTTP, false, stopwatch.Elapsed.TotalMilliseconds, string.Empty, responseExpected: false);
         }
 
         /// <summary>
@@ -532,13 +555,13 @@
             stopwatch.Start();
             DependencyTelemetry operationReturned = (DependencyTelemetry)this.httpProcessingProfiler.OnBeginForBeginGetResponse(request, null, null);
             Thread.Sleep(this.sleepTimeMsecBetweenBeginAndEnd);
-            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");         
+            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");
             Exception exc = new Exception();
             this.httpProcessingProfiler.OnExceptionForEndGetResponse(operationReturned, exc, request, null);
             stopwatch.Stop();
 
             Assert.AreEqual(1, this.sendItems.Count, "Only one telemetry item should be sent");
-            ValidateTelemetryPacket(this.sendItems[0] as DependencyTelemetry, this.testUrl, RemoteDependencyConstants.HTTP, false, stopwatch.Elapsed.TotalMilliseconds, string.Empty);
+            this.ValidateTelemetryPacket(this.sendItems[0] as DependencyTelemetry, this.testUrl, RemoteDependencyConstants.HTTP, false, stopwatch.Elapsed.TotalMilliseconds, string.Empty, responseExpected: false);
         }
 
         #endregion //BeginGetResponse-EndGetResponse
@@ -557,7 +580,7 @@
             var request = WebRequest.Create(this.testUrl);
             DependencyTelemetry operationReturned = (DependencyTelemetry)this.httpProcessingProfiler.OnBeginForBeginGetRequestStream(request, null, null);
             Assert.IsNull(operationReturned, "For async methods, operation returned should be null as correlation is done internally using WeakTables.");
-            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");         
+            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");
         }
 
         /// <summary>
@@ -575,13 +598,13 @@
             stopwatch.Start();
             DependencyTelemetry operationReturned = (DependencyTelemetry)this.httpProcessingProfiler.OnBeginForBeginGetRequestStream(request, null, null);
             Thread.Sleep(this.sleepTimeMsecBetweenBeginAndEnd);
-            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");         
+            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");
             Exception exc = new Exception();
             this.httpProcessingProfiler.OnExceptionForEndGetRequestStream(operationReturned, exc, request, null, null);
             stopwatch.Stop();
 
             Assert.AreEqual(1, this.sendItems.Count, "Only one telemetry item should be sent");
-            ValidateTelemetryPacket(this.sendItems[0] as DependencyTelemetry, this.testUrl, RemoteDependencyConstants.HTTP, false, stopwatch.Elapsed.TotalMilliseconds, string.Empty);
+            this.ValidateTelemetryPacket(this.sendItems[0] as DependencyTelemetry, this.testUrl, RemoteDependencyConstants.HTTP, false, stopwatch.Elapsed.TotalMilliseconds, string.Empty, responseExpected: false);
         }
 
         #endregion //BeginGetRequestStream-EndGetRequestStream
@@ -609,13 +632,13 @@
             Stopwatch stopwatch = new Stopwatch();
             stopwatch.Start();
             this.httpProcessingProfiler.OnBeginForGetRequestStream(request, null);
-            this.httpProcessingProfiler.OnBeginForGetRequestStream(request, null);            
+            this.httpProcessingProfiler.OnBeginForGetRequestStream(request, null);
             Thread.Sleep(this.sleepTimeMsecBetweenBeginAndEnd);
             this.httpProcessingProfiler.OnBeginForGetRequestStream(request, null);
             Thread.Sleep(this.sleepTimeMsecBetweenBeginAndEnd);
 
             this.httpProcessingProfiler.OnBeginForGetResponse(request);
-            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");         
+            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");
             Thread.Sleep(this.sleepTimeMsecBetweenBeginAndEnd);
             this.httpProcessingProfiler.OnEndForGetResponse(null, returnObjectPassed, request);
             stopwatch.Stop();
@@ -625,13 +648,13 @@
             Thread.Sleep(this.sleepTimeMsecBetweenBeginAndEnd);
 
             Assert.AreEqual(1, this.sendItems.Count, "Exactly one telemetry item should be sent");
-            ValidateTelemetryPacket(this.sendItems[0] as DependencyTelemetry, this.testUrl, RemoteDependencyConstants.HTTP, true, stopwatch.Elapsed.TotalMilliseconds, "200");
+            this.ValidateTelemetryPacket(this.sendItems[0] as DependencyTelemetry, this.testUrl, RemoteDependencyConstants.HTTP, true, stopwatch.Elapsed.TotalMilliseconds, "200");
         }
 
-        /// <summary>        
+        /// <summary>
         /// Validates that HttpProcessingProfiler will sent RDD telemetry when GetRequestStream fails and GetResponse is not invoked
         /// 1.create request
-        /// 2.request.GetRequestStream  fails.                
+        /// 2.request.GetRequestStream  fails.
         /// </summary>
         [TestMethod]
         [Description("Validates that HttpProcessingProfiler will sent RDD telemetry when GetRequestStream fails and GetResponse is not invoked.")]
@@ -642,11 +665,11 @@
             var request = WebRequest.Create(this.testUrl);
             
             this.httpProcessingProfiler.OnBeginForGetRequestStream(request, null);
-            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");         
+            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");
             this.httpProcessingProfiler.OnExceptionForGetRequestStream(null, this.ex, request, null);
 
             Assert.AreEqual(1, this.sendItems.Count, "Exactly one telemetry item should be sent");
-            ValidateTelemetryPacket(this.sendItems[0] as DependencyTelemetry, this.testUrl, RemoteDependencyConstants.HTTP, false, 0, string.Empty);
+            this.ValidateTelemetryPacket(this.sendItems[0] as DependencyTelemetry, this.testUrl, RemoteDependencyConstants.HTTP, false, 0, string.Empty, responseExpected: false);
         }
         #endregion //SyncScenarios
 
@@ -677,13 +700,13 @@
             Thread.Sleep(this.sleepTimeMsecBetweenBeginAndEnd);
             this.httpProcessingProfiler.OnBeginForBeginGetResponse(request, null, null);
             Thread.Sleep(this.sleepTimeMsecBetweenBeginAndEnd);
-            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");         
+            Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");
             this.httpProcessingProfiler.OnEndForEndGetResponse(null, returnObjectPassed, request, null);
             stopwatch.Stop();
 
             Assert.AreEqual(1, this.sendItems.Count, "Exactly one telemetry item should be sent");
-            ValidateTelemetryPacket(this.sendItems[0] as DependencyTelemetry, this.testUrl, RemoteDependencyConstants.HTTP, true, stopwatch.Elapsed.TotalMilliseconds, "200");
-        }        
+            this.ValidateTelemetryPacket(this.sendItems[0] as DependencyTelemetry, this.testUrl, RemoteDependencyConstants.HTTP, true, stopwatch.Elapsed.TotalMilliseconds, "200");
+        }
 
         #endregion AsyncScenarios
 
@@ -812,7 +835,7 @@
             Assert.AreEqual(expectedTarget, receivedItem.Target, "HttpProcessingProfiler returned incorrect target for non standard port.");
         }
 
-        #endregion //Misc       
+        #endregion //Misc
 
         #region LoggingTests
 
@@ -835,7 +858,7 @@
                 DependencyTelemetry operationReturned = (DependencyTelemetry)this.httpProcessingProfiler.OnBeginForGetResponse(request);
                 
                 Assert.AreEqual(0, this.sendItems.Count, "No telemetry item should be processed without calling End");
-                TestUtils.ValidateEventLogMessage(listener, "UnexpectedCallbackParameter", EventLevel.Warning);                
+                TestUtils.ValidateEventLogMessage(listener, "UnexpectedCallbackParameter", EventLevel.Warning);
             }
         }
 
@@ -851,8 +874,14 @@
 
         #region Helpers
 
-        private static void ValidateTelemetryPacket(
-            DependencyTelemetry remoteDependencyTelemetryActual, Uri uri, string type, bool success, double expectedValue, string resultCode)
+        private void ValidateTelemetryPacket(
+            DependencyTelemetry remoteDependencyTelemetryActual,
+            Uri uri,
+            string type,
+            bool success,
+            double expectedValue,
+            string resultCode,
+            bool responseExpected = true)
         {
             Assert.AreEqual("GET " + uri.AbsolutePath, remoteDependencyTelemetryActual.Name, true, "Resource name in the sent telemetry is wrong");
             Assert.AreEqual(uri.Host, remoteDependencyTelemetryActual.Target, true, "Resource target in the sent telemetry is wrong");
@@ -860,6 +889,18 @@
             Assert.AreEqual(type.ToString(), remoteDependencyTelemetryActual.Type, "DependencyKind in the sent telemetry is wrong");
             Assert.AreEqual(success, remoteDependencyTelemetryActual.Success, "Success in the sent telemetry is wrong");
             Assert.AreEqual(resultCode, remoteDependencyTelemetryActual.ResultCode, "ResultCode in the sent telemetry is wrong");
+
+            // Validate the http request is present
+            Assert.IsNotNull(this.request, "Http request was not found within the operation details.");
+            Assert.IsNotNull(this.request as WebRequest, "Http request was not the expected type.");
+
+            // If expected -- validate the response
+            if (responseExpected)
+            {
+                Assert.IsNotNull(this.response, "Http response was not found within the operation details.");
+                Assert.IsNotNull(this.response as HttpWebResponse, "Http response was not the expected type.");
+                Assert.IsNull(this.responseHeaders, "Http response headers were not found within the operation details.");
+            }
 
             var valueMinRelaxed = expectedValue - TimeAccuracyMilliseconds;
             Assert.IsTrue(