+++ Src/DependencyCollector/Shared/Implementation/HttpProcessing.cs
@@ -72,7 +72,7 @@ namespace Microsoft.ApplicationInsights.DependencyCollector.Implementation
         internal object OnBegin(object thisObj, bool injectCorrelationHeaders = true)
         {
             try
-            {                
+            {
                 if (thisObj == null)
                 {
                     DependencyCollectorEventSource.Log.NotExpectedCallback(0, "OnBeginHttp", "thisObj == null");
@@ -144,6 +144,7 @@ namespace Microsoft.ApplicationInsights.DependencyCollector.Implementation
                 telemetry.Target = DependencyTargetNameHelper.GetDependencyTargetName(url);
                 telemetry.Type = RemoteDependencyConstants.HTTP;
                 telemetry.Data = url.OriginalString;
+                telemetry.SetOperationDetail(RemoteDependencyConstants.HttpRequestOperationDetailName, webRequest);
 
                 // Add the source instrumentation key header if collection is enabled, the request host is not in the excluded list and the same header doesn't already exist
                 if (this.setCorrelationHeaders && !this.correlationDomainExclusionList.Contains(url.Host))
@@ -215,9 +216,9 @@ namespace Microsoft.ApplicationInsights.DependencyCollector.Implementation
 
         /// <summary>
         /// Common helper for all End Callbacks.
-        /// </summary>        
+        /// </summary>
         /// <param name="request">The HttpWebRequest instance.</param>
-        /// <param name="response">The HttpWebResponse instance.</param>                
+        /// <param name="response">The HttpWebResponse instance.</param>
         internal void OnEndResponse(object request, object response)
         {
             try
@@ -232,6 +233,9 @@ namespace Microsoft.ApplicationInsights.DependencyCollector.Implementation
                         {
                             statusCode = (int)responseObj.StatusCode;
                             this.SetTarget(telemetry, responseObj.Headers);
+
+                            // Set the operation details for the response
+                            telemetry.SetOperationDetail(RemoteDependencyConstants.HttpResponseOperationDetailName, responseObj);
                         }
                         catch (ObjectDisposedException)
                         {
@@ -253,9 +257,9 @@ namespace Microsoft.ApplicationInsights.DependencyCollector.Implementation
 
         /// <summary>
         /// Common helper for all End Callbacks.
-        /// </summary>        
+        /// </summary>
         /// <param name="exception">The exception object if any.</param>
-        /// <param name="request">HttpWebRequest instance.</param>                
+        /// <param name="request">HttpWebRequest instance.</param>
         internal void OnEndException(object exception, object request)
         {
             try
@@ -272,6 +276,9 @@ namespace Microsoft.ApplicationInsights.DependencyCollector.Implementation
                         {
                             statusCode = (int)responseObj.StatusCode;
                             this.SetTarget(telemetry, responseObj.Headers);
+
+                            // Set the operation details for the response
+                            telemetry.SetOperationDetail(RemoteDependencyConstants.HttpResponseOperationDetailName, responseObj);
                         }
                         catch (ObjectDisposedException)
                         {
@@ -302,9 +309,9 @@ namespace Microsoft.ApplicationInsights.DependencyCollector.Implementation
 
         /// <summary>
         /// Common helper for all End Callbacks.
-        /// </summary>        
+        /// </summary>
         /// <param name="request">WebRequest object.</param>
-        /// <param name="statusCode">HttpStatusCode from response.</param>                
+        /// <param name="statusCode">HttpStatusCode from response.</param>
         /// <param name="responseHeaders">Response headers.</param>
         internal void OnEndResponse(object request, object statusCode, object responseHeaders)
         {
@@ -318,6 +325,7 @@ namespace Microsoft.ApplicationInsights.DependencyCollector.Implementation
                     }
 
                     this.SetTarget(telemetry, (WebHeaderCollection)responseHeaders);
+                    telemetry.SetOperationDetail(RemoteDependencyConstants.HttpResponseHeadersOperationDetailName, responseHeaders);
 
                     ClientServerDependencyTracker.EndTracking(this.telemetryClient, telemetry);
                 }