+++ Src/DependencyCollector/Shared/Implementation/ProfilerSqlProcessingBase.cs
@@ -245,6 +245,7 @@
                 telemetry.Type = RemoteDependencyConstants.SQL;
                 telemetry.Target = this.GetDependencyTarget(thisObj);
                 telemetry.Data = commandText;
+                telemetry.SetOperationDetail(RemoteDependencyConstants.SqlCommandOperationDetailName, thisObj);
 
                 // We use weaktables to store the thisObj for correlating begin with end call.
                 this.TelemetryTable.Store(thisObj, new Tuple<DependencyTelemetry, bool>(telemetry, isCustomCreated));
@@ -299,7 +300,7 @@
                         DependencyCollectorEventSource.Log.CallbackError(thisObj == null ? 0 : thisObj.GetHashCode(), "OnEndAsyncSql", ex);
                     }
                 });
-            }            
+            }
             catch (Exception ex)
             {
                 DependencyCollectorEventSource.Log.CallbackError(thisObj == null ? 0 : thisObj.GetHashCode(), "OnEndAsyncSql", ex);